Add documentation setup (#4)
* Add Documentation setup with Documenter

This adds no actual documentation, only the infrastructure needed to
set up and deploy documentation.

* Remove testing on 0.7, add 1.1

In the REQUIRE file, the package declares itself to be compatible only
with Julia 1.0 or higher, but 0.7 was still listed in CI. This removes
0.7 and adds 1.1.

It also updates the coverage setup.

* Add a Project.toml file

These are indispensible for local development, and will soon be required
anyway (for some definition of "soon").

* Add missing dependency on ForwardDiff to REQUIRE