auto-cancel old PR builds
We are beginning to hit concurrency limits of GH actions if there are
many pushes in a row since we do run quite a lot of jobs. This should
automatically cancel the old jobs for a particular PR if there is a more
recent push.

All credit goes to @DilumAluthge.