Remove major invalidations from ==
See https://github.com/JuliaDiff/ChainRulesCore.jl/issues/523 . Seems to only be used for tests according to the PR but it's a major compile time problem, so it doesn't make sense to keep them. This can be marked breaking if it breaks downstream, but it's rather easy to update downstream AD packages for this so the value proposition is pretty clear.